requirements.txt
netmiko==4.3.0
napalm==4.1.0
requests==2.31.0
prometheus-client==0.19.0
pyyaml==6.0.1
rich==13.7.0
